pyproject.toml
[project]
name = "rockefeller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.160.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "ruff>=0.9.3",
    "typer>=0.15.1",
]
authors = [
    {name = "Omar Bohsali", email = "omar.bohsali@gmail.com"},
]
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"
# Add these lines:
lint.select = ["I"]
fix = true

[lint.isort]
known-first-party = ["rockefeller"]
